HPCC-12127 Use 64-bit integer for HTTP Content-Length header
In the existing ESP code, two member variables are used to store
and set HTTP Content-Length header: m_content_length64 (int64)
and m_content_length (int). When an HTTP message content is very
large, m_content_length may have an incorrect value. ESP fails
to download a >4G file due to the incorrect value. This fix
changes the m_content_length to be int64 and replaces the
m_content_length64 with the m_content_length.

+++ esp/bindings/http/platform/mime.hpp
@@ -80,7 +80,7 @@ public:
     void readUploadFileName(MemoryBuffer& fileContent, StringBuffer& fileName);
 
     void serialize(StringBuffer& contenttype, StringBuffer & buffer);
-    void unserialize(const char* contenttype, int text_length, const char* text);
+    void unserialize(const char* contenttype, __int64 text_length, const char* text);
 };
 
 #endif
